Testes

+++ src/main/java/org/example/order/application/controllers/create/requests/PedidoCreateItemRequest.java
@@ -1,7 +1,5 @@
 package org.example.order.application.controllers.create.requests;
 
-import java.math.BigDecimal;
-
-public record PedidoCreateItemRequest(String produtoNome, BigDecimal valor, int quantidade) {
+public record PedidoCreateItemRequest(String produtoNome, double valor, int quantidade) {
 
 }
