+++ src/main/java/org/example/order/core/applications/usecases/criarPedido/CriarPedidoItemInput.java
@@ -1,6 +1,4 @@
 package org.example.order.core.applications.usecases.criarPedido;
 
-import java.math.BigDecimal;
-
-public record CriarPedidoItemInput(String produtoNome, BigDecimal valor, int quantidade) {
+public record CriarPedidoItemInput(String produtoNome, double valor, int quantidade) {
 }
